refactor: Move files to subfolders
