+++ src/devices/gps_device.cpp
