+++ src/devices/serial_gps_device.h
